pyproject.toml
[project]
name = "fastmcp-agent-framework"
version = "0.1.0"
description = "Generic FastMCP Agent Framework with LLM integration for building intelligent tool-enabled agents"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "asyncio-mqtt>=0.16.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
fastmcp-agent = "fastmcp_agent.main:main"
calculator-agent = "fastmcp_agent.examples.calculator:main_sync"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fastmcp_agent"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 